Made small changes to file opening and linting
Opening file now accepts a file path as an optional parameter; if set, the IDE will open that file instead of asking to open. This will still allow the open file to be run as a shortcut later.

Linting now only occurs on save, need to work on that one before it is fully functional.

Now opens last opened file, will become more useful when multiple files can be opened at once.

Added sample project to work on as a testing field. These files will be subject to change as I test.

There is now no lint key shortcut.

Signed-off-by: Keith Allatt <keith.allatt@gmail.com>